SWARM-276 - wildfly-swarm:stop doesn't work from the CLI
The resolution was to simply provide an error message and continue
to not work from the CLI.

When used as part of integration tests, the :start and :stop Mojos
can communicate objects to each other.

From the command-line, while we could write a PID file and use that
during a subsequent mvn wildfly-swarm:stop CLI invokation, it is
non-portable.  Plus Java doesn't have a decent way to send signals
to other processes.

jnr-posix or some other library may be able to help us out here,
but I'll file that enhancement under an additional issue.