Pull fraction package specs from FractionList
Motivation:
For fraction auto-detection, we maintain a map of fraction names to a
spec of packages that indicate the fraction should be used. Now that
fraction versions come from the BOM, we need to pull the package info
from something that has access to the BOM, so this modifies FractionList
to provide the package specs as well.

Modifications:
* create an interface for the FractionList that resides in this project
* move FractionDescriptor from -swarm to here so we can use it in both
  projects
* allow passing a FractionList to BuildTool

Result:
Fraction detection now uses the correct versions for fractions.