Don't use internal aether class
Motivation:
The maven provided by fedora can't inject a DefaultRepositorySystem, so
can't be used with wildfly-swarm-plugin. DefaultRepositorySystem is an
internal class, so we shouldn't be using it. Asking for the public
interface (RepositorySystem) works, so we'll do that instead.

Modifications:
* loosen the type of the repositorySystem

Result:
The fedora-provided maven can now be used to build swarm apps.