Pass through defined repos to jboss-modules [SWARM-337]
jboss-modules resolves g:a:v's from module.xml files at runtime. If the
artifact isn't in M2_REPO, it can resolve remotely, but, by default,
only resolves against central. You can provide additional repos, so this
provides the repos set in the pom.

If a user hasn't build swarm locally, the -runtime deps won't be in
M2_REPO, so need to be resolved remotely, which this should allow.
However, jboss-modules doesn't properly handle resolving remote
SNAPSHOTs, so this is but the first step towards a real solution for
SWARM-337.