Subject: Older Fraction versions selected
Motivation:
SWARM-398, when auto detection is used to determine the required WildFly Swarm dependencies it would correctly identify the required dependencies but transitive dependencies could result in older versions being selected.

Modifications:
Create a VersionSelector for Maven that resolves a conflict by choosing the newest version of a dependency, then use this selector when collecting all the dependencies that are transitive from those that have been auto detected.

Result:
Will prevent an older version of a WildFly Swarm dependency, that's transitive, from being selected when auto detection is used in a user application