Sleeping in shutdown hook will give time to shutdown gracefully
Motivation:
This fixes SWARM-404. As Maven is the main process when running
wildfly-swarm:run, it's important to give time to shutdown before
attempting to destroy the process completely.

Modifications:
A Thread.sleep(100L) was added to the shutdown hook so the app can
shutdown gracefully

Result:
The WildFly Core 2.0.10.Final "Kenny" stopped message appears when
the app is interrupted.