Move snapshot bom deployment to a profile
When releasing, we want only the release bom deployment to run (in the
'release' profile), so this moves the snapshot deployment (that happens
on CI) to a profile that is active by default. According to
https://maven.apache.org/guides/introduction/introduction-to-profiles.html,
activeByDefault profiles are automatically deactivated when any other
profile in the pom is activated. In this case, the only other profile is
'release'.