SWARM-434: Support executable jars
Motivation: The reason for having executable jars lies on making it possible
to be easily started as Unix/Linux services using either init.d or systemd.

Result: The PackageMojo and PackageTask now declares an executable property that prepends a sh script
and sets the JAR to be a an executable.